feat: implement MVP with Wayland connectivity and TEA engine
   - Generate river-window-management-v1 bindings via wayland-nim
   - Implement main event loop with Registry and Manager binding
   - Establish TEA (Model-Update-View) message processing loop
   - Port Mango 'Column of Stacks' scroller logic to src/layouts/scroller.nim
   - Setup .gitignore and project structure

+++ src/layouts/scroller.nim
@@ -34,7 +34,7 @@ proc layoutScroller*(tag: TagState, screen: Rect, outerGap, innerGap: int32): se
     
     for winId in col.windows:
       # For now, equal height proportions if not specified
-      let winHeight = usableColHeight div int32(numWindows)
+      let winHeight = int32(usableColHeight div int32(numWindows))
       
       instructions.add(RenderInstruction(
         windowId: winId,
